# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ilm School Portal API
  version: 1.0.0
  description: |-
    Multi-school (multi-tenant) school management API. Every role except
    `system_admin` is scoped to exactly one school (`schoolId` on the JWT);
    `system_admin` is a platform-level role with no school of its own and
    can act across every school.

    **Auth**: session is an httpOnly `token` cookie (JWT), set by
    `POST /auth/login` and cleared by `POST /auth/logout`. There is no
    bearer-token flow — the cookie is sent automatically by the browser on
    same-origin requests (`withCredentials: true` on the frontend's axios
    client). A request with no cookie, an invalid/expired token, or a
    revoked token (see below) gets `401`.

    **Session lifetime**: tokens are signed for 1 hour. Any authenticated
    request past the halfway point of that window silently reissues the
    cookie, so an active user is never hard-logged-out mid-session — only
    genuine inactivity lets it expire. That sliding renewal is itself capped
    at 12 hours from the original login (`sessionStartedAt` in the token),
    so a session can't be kept alive indefinitely just by staying active —
    after 12 hours the user has to log in again regardless of activity.
    Logging out invalidates the token server-side immediately (via a
    `token_version` check on every request), not just client-side — and
    because there's no per-device session tracking, logging out on one
...
    sessions.

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /parent/add-child:
    post:
      tags:
        - Parent
      summary: Add a child to the logged-in parent's account and enroll them in a class
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - surname
                - gender
                - date_of_birth
                - class_code
              properties:
                first_name:
                  type: string
                middle_name:
                  type: string
                surname:
                  type: string
                gender:
                  type: string
                  enum:
                    - male
                    - female
                date_of_birth:
                  type: string
                  format: date
                class_code:
                  type: string
                  description: Must exist within the parent's own school
                address1:
                  type: string
                address2:
                  type: string
                address3:
                  type: string
                city:
                  type: string
                postcode:
                  type: string
                medical_condition:
                  type: string
      responses:
        '200':
          description: Child added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  children:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChildSummary'
        '400':
          description: 'Missing required fields, or invalid class code'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postParentAddChild
  /parent/link-guardian:
    post:
      tags:
        - Parent
      summary: Self-service request to link as an additional guardian on an existing child
      description: For an already-approved, logged-in parent linking to a child that isn't theirs yet (e.g. added later, not at initial registration). Always inserted `pending` — there's no accompanying user-approval action for this to ride along with, so it always needs an explicit admin decision via GET/POST /admin/guardian-requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - guardian_code
              properties:
                guardian_code:
                  type: string
      responses:
        '201':
          description: 'Request submitted, pending admin approval'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Parent not found, or guardian_code missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invalid guardian code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Already linked, or a request is already pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postParentLinkGuardian
...
  /parent/schedule:
    get:
      tags:
        - Parent
      summary: Class schedule for all of the parent's children, grouped by child
      description: Gated by `timetable`. Each child's slots come from whichever class(es) they're enrolled in via `student_classes`; a child with no enrolled class or no slots simply won't appear.
      responses:
        '200':
          description: Schedule, grouped by child
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedule:
                    type: array
                    items:
                      type: object
                      properties:
                        student_id:
                          type: integer
                        child_name:
                          type: string
                        slots:
                          type: array
                          items:
                            type: object
                            properties:
                              slot_date:
                                type: string
                                format: date
                              start_time:
                                type: string
                              end_time:
                                type: string
                              class_name:
                                type: string
                              subject_name:
                                type: string
                                nullable: true
                              teacher_first_name:
                                type: string
                              teacher_surname:
                                type: string
        '403':
          description: The timetable feature is disabled for this school
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getParentSchedule
...
  /parent/direct-debit/mandate:
    get:
      tags:
        - Parent
      summary: Get the caller's own direct-debit mandate status
      description: Parent only. Gated by `direct_debit`.
      responses:
        '200':
          description: The parent's mandate, or null if never set up
          content:
            application/json:
              schema:
                type: object
                properties:
                  mandate:
                    type: object
                    nullable: true
                    properties:
                      status:
                        type: string
                        enum: [pending, active, cancelled]
                      created_at:
                        type: string
                        format: date-time
                      cancelled_at:
                        type: string
                        format: date-time
                        nullable: true
        '403':
          $ref: '#/components/responses/Forbidden'
      operationId: getParentDirectDebitMandate
    post:
      tags:
        - Parent
      summary: Self-serve set up (or re-set-up) a direct-debit mandate
      description: |-
        Parent only. Gated by `direct_debit`. Creates the mandate via the
        configured payment provider (currently a stub — no real bank
        details ever touch this app) and stores it active. If the parent
        previously cancelled a mandate, this reactivates the same row with
        fresh provider IDs rather than creating a second one — there is
        only ever one mandate per parent (`payment_mandates.parent_id` is
        unique).
      responses:
        '200':
          description: An existing cancelled mandate was reactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '201':
          description: A new mandate was created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A mandate is already active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postParentDirectDebitMandate
  /parent/direct-debit/mandate/cancel:
    post:
      tags:
        - Parent
      summary: Cancel the caller's own direct-debit mandate
      description: Parent only. Gated by `direct_debit`. Does not affect fees already `pending_collection` under the mandate — those still settle via the provider webhook.
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No active mandate to cancel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postParentDirectDebitMandateCancel
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly JWT set by POST /auth/login. See the top-level description for session lifetime/refresh/revocation behavior.
...
  responses:
    MessageOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
    Forbidden:
      description: 'Authenticated, but not permitted (wrong role, or outside the caller''s school)'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ChildSummary:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        surname:
          type: string
        gender:
          type: string
        date_of_birth:
          type: string
          format: date
        guardian_code:
          type: string
          description: Share this with a second guardian so they can self-request linking via POST /parent/link-guardian
...